added user validation before editing menu or restaurant item

+++ controllers/menu_item_cont.py
@@ -1,9 +1,17 @@
 from flask import Blueprint, flash, render_template, url_for, request, redirect
 from models.menu_items import MenuItem, Restaurant, db
 from flask import session as login_session
+from controllers.restaurant_cont import validate_owner
 
 # create a blueprint for all menu items
 menu_items = Blueprint("menu_item", __name__)
+
+
+def validate_menu_owner(user_id, menu_item):
+    if user_id == menu_item.restaurant.user_id:
+        return True
+    else:
+        return False
 
 
 # route to create a menu item for a given restaurant
@@ -12,20 +20,28 @@ menu_items = Blueprint("menu_item", __name__)
 @menu_items.route("/restaurant/<int:restaurant_id>/menu/new", methods=['GET', 'POST'])
 def create(restaurant_id):
     user_id = login_session.get('id')
-    restaurant = Restaurant.query.get(restaurant_id)
     if user_id:
-        if request.method == 'POST':
-            restaurant = Restaurant.query.get(restaurant_id)
-            name = request.form.get("name")
-            desc = request.form.get("desc")
-            price = request.form.get("price")
-            menu_item = MenuItem(name=name, desc=desc, price=price, restaurant_id=restaurant.id)
-            db.session.add(menu_item)
-            db.session.commit()
-            flash("Menu item added successfully!")
-            return redirect(url_for("restaurant.show_menu", restaurant_id=restaurant.id))
+        restaurant = Restaurant.query.get(restaurant_id)
+        if restaurant:
+            if validate_owner(user_id, restaurant):
+                if request.method == 'POST':
+                    restaurant = Restaurant.query.get(restaurant_id)
+                    name = request.form.get("name")
+                    desc = request.form.get("desc")
+                    price = request.form.get("price")
+                    menu_item = MenuItem(name=name, desc=desc, price=price, restaurant_id=restaurant.id)
+                    db.session.add(menu_item)
+                    db.session.commit()
+                    flash("Menu item added successfully!")
+                    return redirect(url_for("restaurant.show_menu", restaurant_id=restaurant.id))
+                else:
+                    return render_template("menu/form.html", menu_item='', restaurant=restaurant)
+            else:
+                flash("you can not edit this menu as you are not the owner")
+                return redirect(url_for('restaurant.index'))
         else:
-            return render_template("menu/form.html", menu_item='', restaurant=restaurant)
+            flash("restaurant menu not found please try again")
+            return redirect(url_for('restaurant.index'))
     else:
         flash('Operation not allowed. Please log in to create a menu item')
         return redirect(url_for('restaurant.index'))
@@ -43,19 +59,23 @@ def update(restaurant_id, menu_item_id):
         restaurant = Restaurant.query.get(restaurant_id)
         menu_item = MenuItem.query.get(menu_item_id)
         if menu_item:
-            if request.method == 'GET':
-                return render_template("menu/form.html", menu_item=menu_item, restaurant=restaurant)
-            elif request.method == 'POST':
-                menu_item.name = request.form.get("name")
-                menu_item.desc = request.form.get("desc")
-                menu_item.price = request.form.get("price")
-                db.session.add(menu_item)
-                db.session.commit()
-                flash("Menu item updated successfully!")
-                return redirect(url_for("restaurant.show_menu", restaurant_id=menu_item.restaurant.id))
+            if validate_menu_owner(user_id, menu_item):
+                if request.method == 'GET':
+                    return render_template("menu/form.html", menu_item=menu_item, restaurant=restaurant)
+                elif request.method == 'POST':
+                    menu_item.name = request.form.get("name")
+                    menu_item.desc = request.form.get("desc")
+                    menu_item.price = request.form.get("price")
+                    db.session.add(menu_item)
+                    db.session.commit()
+                    flash("Menu item updated successfully!")
+                    return redirect(url_for("restaurant.show_menu", restaurant_id=menu_item.restaurant.id))
+                else:
+                    flash("Invalid Action, Please try again")
+                    return redirect(url_for("restaurant.menu", restaurant_id=menu_item.restaurant.id))
             else:
-                flash("Invalid Action, Please try again")
-                return redirect(url_for("restaurant.menu", restaurant_id=menu_item.restaurant.id))
+                flash("you can not edit this restaurant as you are not the owner")
+                return redirect(url_for('restaurant.index'))
         else:
             flash("Item not found, please try again")
             return redirect(url_for("restaurant.index"))
@@ -75,16 +95,20 @@ def delete(restaurant_id, menu_item_id):
         restaurant = Restaurant.query.get(restaurant_id)
         menu_item = MenuItem.query.get(menu_item_id)
         if menu_item:
-            if request.method == "GET":
-                return render_template("menu/delete.html", menu_item=menu_item)
-            elif request.method == "POST":
-                db.session.delete(menu_item)
-                db.session.commit()
-                flash("Menu item successfully deleted!")
-                return redirect(url_for("restaurant.show_menu", restaurant_id=restaurant.id))
+            if validate_menu_owner(user_id, menu_item):
+                if request.method == "GET":
+                    return render_template("menu/delete.html", menu_item=menu_item)
+                elif request.method == "POST":
+                    db.session.delete(menu_item)
+                    db.session.commit()
+                    flash("Menu item successfully deleted!")
+                    return redirect(url_for("restaurant.show_menu", restaurant_id=restaurant.id))
+                else:
+                    flash("Invalid operation. Please try again")
+                    return redirect(url_for("restaurant.show_menu", restaurant_id=restaurant.id))
             else:
-                flash("Invalid operation. Please try again")
-                return redirect(url_for("restaurant.show_menu", restaurant_id=restaurant.id))
+                flash("you can not edit this restaurant as you are not the owner")
+                return redirect(url_for('restaurant.index'))
         else:
             flash("Menu item not found. Please try again")
             return redirect(url_for("restaurant.show_menu", restaurant_id=restaurant.id))
